+++ src/Symfony/Component/Cache/Traits/FilesystemCommonTrait.php
@@ -26,7 +26,7 @@ trait FilesystemCommonTrait
     private function init(string $namespace, ?string $directory)
     {
         if (!isset($directory[0])) {
-            $directory = sys_get_temp_dir().'/symfony-cache';
+            $directory = sys_get_temp_dir().\DIRECTORY_SEPARATOR.'symfony-cache';
         } else {
             $directory = realpath($directory) ?: $directory;
         }
@@ -55,12 +55,12 @@ trait FilesystemCommonTrait
     {
         $ok = true;
 
-        foreach (new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($this->directory, \FilesystemIterator::SKIP_DOTS)) as $file) {
+        foreach ($this->scanHashDir($this->directory) as $file) {
             if ('' !== $namespace && 0 !== strpos($this->getFileKey($file), $namespace)) {
                 continue;
             }
 
-            $ok = ($file->isDir() || $this->doUnlink($file) || !file_exists($file)) && $ok;
+            $ok = ($this->doUnlink($file) || !file_exists($file)) && $ok;
         }
 
         return $ok;
@@ -121,6 +121,33 @@ trait FilesystemCommonTrait
     private function getFileKey(string $file): string
     {
         return '';
+    }
+
+    private function scanHashDir(string $directory): \Generator
+    {
+        if (!file_exists($directory)) {
+            return;
+        }
+
+        $chars = '+-ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
+
+        for ($i = 0; $i < 38; ++$i) {
+            if (!file_exists($directory.$chars[$i])) {
+                continue;
+            }
+
+            for ($j = 0; $j < 38; ++$j) {
+                if (!file_exists($dir = $directory.$chars[$i].\DIRECTORY_SEPARATOR.$chars[$j])) {
+                    continue;
+                }
+
+                foreach (@scandir($dir, SCANDIR_SORT_NONE) ?: [] as $file) {
+                    if ('.' !== $file && '..' !== $file) {
+                        yield $dir.\DIRECTORY_SEPARATOR.$file;
+                    }
+                }
+            }
+        }
     }
 
     /**