+++ src/Symfony/Component/Cache/Traits/FilesystemTrait.php
@@ -33,7 +33,7 @@ trait FilesystemTrait
         $time = time();
         $pruned = true;
 
-        foreach (new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($this->directory, \FilesystemIterator::SKIP_DOTS), \RecursiveIteratorIterator::LEAVES_ONLY) as $file) {
+        foreach ($this->scanHashDir($this->directory) as $file) {
             if (!$h = @fopen($file, 'rb')) {
                 continue;
             }